Generalize `generate_model_outputs_using_core_keras_layers()`.
This change adds the following three new features to the above function:
(i) it supports nested custom layers of depth >2;
(ii) it supports custom layers with sublayers of the same name;
(iii) it allows the caller to exclude certain layers from the expansion.

Features (ii)-(iii) will be needed for the development of DP models that use
Trasformer or BERT-type layers.

PiperOrigin-RevId: 516841817